amended votes - moved to SingleArticle only. Moved vote state from app down to votes. Changed CSS for mobile - font size lowered
