+++ src/App.css
@@ -11,6 +11,7 @@
 body {
   background-color: floralwhite;
   margin-top: 1rem;
+  font-size: 16px;
 }
 
 header h1 {
@@ -56,7 +57,7 @@ nav {
   justify-content: center;
 }
 
-.articles-container-card  {
+.articles-container-card {
   display: flex;
   flex-direction: column;
   margin: 1rem;
@@ -91,6 +92,11 @@ nav {
   text-align: center;
 }
 
+.votes-error {
+  width: auto;
+  text-align: center;
+}
+
 .votes-buttons {
   display: flex;
   flex-direction: row;
@@ -102,7 +108,7 @@ nav {
 .votes-buttons button {
   width: 2rem;
   height: 1.5rem;
-  background-color:rgb(3, 3, 2);
+  background-color: purple;
   border-radius: 1rem;
   border: 1px solid;
 }
@@ -153,7 +159,11 @@ nav {
   padding: 0.5rem;
 }
 
-@media (max-width: 650px) {
+@media screen and (max-width: 600px) {
+
+  body {
+    font-size: 10px;
+  }
 
   .comments-container {
     width: 80vw;
@@ -161,6 +171,10 @@ nav {
 
   .comments-container-card {
     width: 90%;
+  }
+
+  .votes-error {
+    width: 40%;
   }
 
 }
