+++ src/components/SingleArticle/Votes.jsx
@@ -1,7 +1,9 @@
 import { patchVote } from "../../utils/api"
+import { useState } from "react"
 
-export default function Votes({ vote, setVote, article_id, votes }) {
+export default function Votes({ article_id, votes }) {
 
+    const [vote, setVote] = useState(0)
     const [error, setError] = useState("")
 
     const updateVoteNum = (num) => {
@@ -9,23 +11,28 @@ export default function Votes({ vote, setVote, article_id, votes }) {
             return currVote + num;
         })
         patchVote(article_id, num).catch((err) => {
+            setError("Please refresh or try again later!")
             setVote((currVote) => {
                 return currVote - num;
             })
-            setError("Something went wrong!")
         })
     }
 
     if (error !== "") {
-        return <p>{error}</p>
+        return (
+            <div className="votes-error">
+                <p>Votes: {votes + vote}</p>
+                <p>{error}</p>
+            </div>
+        )
     }
 
     return (
         <div className="votes">
             <p>Votes: {votes + vote}</p>
             <div className="votes-buttons">
-                <button onClick={() => updateVoteNum(-1)} value="-1">👎</button>
-                <button onClick={() => updateVoteNum(1)} value="1">👍</button>
+                <button onClick={() => updateVoteNum(-1)} disabled={vote === -1}>👎</button>
+                <button onClick={() => updateVoteNum(1)} disabled={vote === 1} >👍</button>
             </div>
         </div>
     )
